Convert to using long double

+++ functions.h
@@ -2,4 +2,4 @@
  * Define the prototypes of functions defined in functions.c
  */
 
-double D(double k_perp, double omega);
+long double D(double k_perp, double omega);
